+++ arangod/Graph/AttributeWeightShortestPathFinder.cpp
@@ -157,7 +157,7 @@ AttributeWeightShortestPathFinder::AttributeWeightShortestPathFinder(ShortestPat
       _highscoreSet(false),
       _highscore(0),
       _bingo(false),
-      _resultCode(TRI_ERROR_NO_ERROR),
+      _resultCode(static_cast<int>(TRI_ERROR_NO_ERROR)),
       _intermediateSet(false),
       _intermediate() {
   // cppcheck-suppress *
@@ -173,7 +173,7 @@ void AttributeWeightShortestPathFinder::clear() {
   _highscoreSet = false;
   _highscore = 0;
   _bingo = false;
-  _resultCode = TRI_ERROR_NO_ERROR;
+  _resultCode = static_cast<int>(TRI_ERROR_NO_ERROR);
   _intermediateSet = false;
   _intermediate = arangodb::velocypack::StringRef{};
 }
